refactor: co-locate scattered config by domain
- Split commands.ts (378 lines) into 7 domain-specific files:
  commonCommands, mainViewCommands, progressViewCommands,
  historyViewCommands, profileViewCommands, memoryViewCommands,
  settingsViewCommands. Original file becomes a re-export shim.

- Consolidate provider config into shared/constants/providers.ts:
  PROVIDER_URLS (from apiKeyCommands.ts),
  PROVIDER_VSCODE_SETTINGS (from SettingsViewMessageHandler.ts)

- Move LaTeX timing constants (DIFF_REGISTRATION_DELAY_MS,
  LATEX_VIEWER_*_DELAY_MS) to shared/constants/latex.ts

- Extract provider streaming/endpoint config into
  utils/config/providerConfig.ts (from utils/config/constants.ts)

- Fix common→agent dependency violation:
  executionRequests.ts → agent/core/executionRequests.ts
  mainViewStateUtils.ts → frontend/mainViewStateUtils.ts
  Old locations become re-export shims for backward compatibility.
