+++ src/common/execution/executionRequests.ts
@@ -1,46 +1,10 @@
-// Local imports - agent config
-import { AgentConfigSchema } from '@agent/core';
-import type { AgentConfig, AgentConfigInput } from '@agent/core/AgentConfig';
-
-// Local imports - shared schemas
-import type { ExecutionId } from '@shared/schemas';
-
-// Type imports
-import type { z } from 'zod';
-
-export interface ExecutionRequest {
-  config: AgentConfigInput;
-  executionId?: ExecutionId;
-}
-
-export interface ValidatedExecutionRequest {
-  config: AgentConfig;
-  executionId?: ExecutionId;
-}
-
-export type ExecutionValidationResult =
-  | { valid: true; request: ValidatedExecutionRequest }
-  | { valid: false; message: string; issue?: z.ZodIssue };
-
-export function validateExecutionRequest(
-  request: ExecutionRequest,
-): ExecutionValidationResult {
-  const parseResult = AgentConfigSchema.safeParse(request.config);
-  if (!parseResult.success) {
-    const issue = parseResult.error.issues[0];
-    const errorPath = issue?.path.join('.') || 'unknown';
-    return {
-      valid: false,
-      message: `Invalid configuration (${errorPath}): ${issue?.message ?? 'validation failed'}`,
-      issue,
-    };
-  }
-
-  return {
-    valid: true,
-    request: {
-      config: parseResult.data,
-      executionId: request.executionId,
-    },
-  };
-}
+/**
+ * Re-export from canonical location.
+ * Moved to @agent/core/executionRequests to fix common→agent dependency violation.
+ */
+export {
+  validateExecutionRequest,
+  type ExecutionRequest,
+  type ValidatedExecutionRequest,
+  type ExecutionValidationResult,
+} from '@agent/core/executionRequests';
