+++ src/common/webview/commands.ts
@@ -1,377 +1,15 @@
 /**
- * Standardized command constants for all webviews
+ * Backward-compatible re-exports of all webview command constants.
+ *
+ * Each command group now lives in its own domain-specific file.
+ * Import directly from the domain file when possible:
+ *   import { MAIN_VIEW_COMMANDS } from '@common/webview/mainViewCommands';
+ *   import { PROGRESS_VIEW_COMMANDS } from '@common/webview/progressViewCommands';
  */
-
-// Common commands used across all views
-export const COMMON_COMMANDS = {
-  THEME_SET: 'setTheme',
-  DEBUG_MODE_SET: 'setDebugMode',
-  STATE_RESTORE: 'restoreState',
-  WEBVIEW_READY: 'webviewReady',
-  ERROR: 'error',
-  SWITCH_VIEW: 'switchView',
-} as const;
-
-// Main view specific commands
-export const MAIN_VIEW_COMMANDS = {
-  ...COMMON_COMMANDS,
-  // File operations
-  FILE_SELECT: 'selectFile',
-  FILE_SELECTED: 'fileSelected',
-
-  // Execution
-  EXECUTE: 'execute',
-  MERGE: 'merge',
-  COMPARE: 'compare',
-
-  // Settings
-  MODEL_SELECTED: 'modelSelected',
-  SETTINGS_OPEN: 'openSettings',
-
-  // File selection cases
-  SELECT_INPUT_FILE: 'selectInputFile',
-  SELECT_REFERENCE_FILE: 'selectReferenceFile',
-  SELECT_AUXILIARY_FILE: 'selectAuxiliaryFile',
-  SELECT_MEDIA_FILE: 'selectMediaFile',
-  SELECT_EDITED_FILE: 'selectEditedFile',
-
-  // File selected cases
-  INPUT_FILE_SELECTED: 'inputFileSelected',
-  REFERENCE_FILE_SELECTED: 'referenceFileSelected',
-  AUXILIARY_FILE_SELECTED: 'auxiliaryFileSelected',
-  MEDIA_FILE_SELECTED: 'mediaFileSelected',
-  EDITED_FILE_SELECTED: 'editedFileSelected',
-
-  // Request file cases
-  REQUEST_INPUT_FILE: 'requestInputFile',
-  REQUEST_REFERENCE_FILE: 'requestReferenceFile',
-  REQUEST_AUXILIARY_FILE: 'requestAuxiliaryFile',
-  REQUEST_MEDIA_FILE: 'requestMediaFile',
-  REQUEST_EDITED_FILE: 'requestEditedFile',
-  REQUEST_BASE_FILE: 'requestBaseFile',
-  REQUEST_DEFAULT_OUTPUT_FILES: 'requestDefaultOutputFiles',
-
-  // Multiple file operations
-  SET_INPUT_FILES: 'setInputFiles',
-  SET_REFERENCE_FILES: 'setReferenceFiles',
-  SET_AUXILIARY_FILES: 'setAuxiliaryFiles',
-  SET_MEDIA_FILES: 'setMediaFiles',
-  SELECT_MULTIPLE_FILES: 'selectMultipleFiles',
-
-  // Other operations
-  SHOW_INFORMATION_MESSAGE: 'showInformationMessage',
-  SHOW_INSTRUCTION: 'showInstruction',
-  GET_THEME: 'getTheme',
-  GET_DEBUG_MODE: 'getDebugMode',
-  GET_CURRENT_FILE: 'getCurrentFile',
-  ADD_OPENED_FILES: 'addOpenedFiles',
-  POLISH_INSTRUCTION_TEXT: 'polishInstructionText',
-  TRANSCRIBE_INSTRUCTION: 'transcribeInstruction',
-  START_RECORDING: 'startRecording',
-  STOP_RECORDING: 'stopRecording',
-  RECORDING_STOPPED: 'recordingStopped',
-  SHOW_AGENT_HISTORY: 'showAgentHistory',
-  OPEN_AGENT_SETTINGS: 'openAgentSettings',
-  OPEN_MODEL_SETTINGS: 'openModelSettings',
-  CLIPBOARD_IMAGE: 'clipboardImage',
-  OPEN_SET_API_KEY: 'openSetApiKey',
-  OPEN_SET_PROVIDER_API_KEY: 'openSetProviderApiKey',
-  OPEN_PROVIDER_API_KEY_URL: 'openProviderApiKeyUrl',
-  OPEN_API_KEY_GUIDE: 'openApiKeyGuide',
-  SHOW_API_KEY_BANNER: 'showApiKeyBanner',
-  HIDE_API_KEY_BANNER: 'hideApiKeyBanner',
-  OPEN_AGENT_DIRECTORY: 'openAgentDirectory',
-  OPEN_AGENT_DOCS: 'openAgentDocs',
-  OPEN_INSTALLATION_DOCS: 'openInstallationDocs',
-  OPEN_INSTALL_GUIDE: 'openInstallGuide',
-  RECHECK_DEPENDENCIES: 'recheckDependencies',
-  SHOW_AGENT_CONFIG_BANNER: 'showAgentConfigBanner',
-  HIDE_AGENT_CONFIG_BANNER: 'hideAgentConfigBanner',
-  SHOW_DEPENDENCY_BANNER: 'showDependencyBanner',
-  HIDE_DEPENDENCY_BANNER: 'hideDependencyBanner',
-  UPDATE_DEPENDENCY_REMINDER_SETTING: 'updateDependencyReminderSetting',
-  SHOW_GETTING_STARTED_BANNER: 'showGettingStartedBanner',
-  HIDE_GETTING_STARTED_BANNER: 'hideGettingStartedBanner',
-  SHOW_LOGIN_BANNER: 'showLoginBanner',
-  HIDE_LOGIN_BANNER: 'hideLoginBanner',
-  SIGN_IN_FROM_BANNER: 'signInFromBanner',
-  DISMISS_LOGIN_BANNER: 'dismissLoginBanner',
-
-  // Extension response events
-  INSTRUCTION_TEXT_POLISHED: 'instructionTextPolished',
-  INSTRUCTION_TEXT_POLISH_ERROR: 'instructionTextPolishError',
-  INSTRUCTION_TEXT_TRANSCRIBED: 'instructionTextTranscribed',
-  RECORDING_STARTED: 'recordingStarted',
-  RECORDING_ERROR: 'recordingError',
-  SET_INPUT_FILE: 'setInputFile',
-  SET_REFERENCE_FILE: 'setReferenceFile',
-  SET_AUXILIARY_FILE: 'setAuxiliaryFile',
-  SET_MEDIA_FILE: 'setMediaFile',
-  SET_EDITED_FILE: 'setEditedFile',
-  SET_DEFAULT_OUTPUT_FILES: 'setDefaultOutputFiles',
-  ADD_MEDIA_FILE: 'addMediaFile',
-  SET_OUTPUT_FILES: 'setOutputFiles',
-  SET_RECENT_COMMITS: 'setRecentCommits',
-  SET_CURRENT_FILE: 'setCurrentFile',
-  SET_OPENED_FILES: 'setOpenedFiles',
-  SET_BASE_FILE: 'setBaseFile',
-  SET_ALL_SINGLE_FILES: 'setAllSingleFiles',
-  SET_SELECTED_COMMIT: 'setSelectedCommit',
-  SET_MODEL_OPTIONS: 'setModelOptions',
-  SET_AGENT_OPTIONS: 'setAgentOptions',
-  SET_SELECTED_AGENT: 'setSelectedAgent',
-
-  // File refresh and update operations
-  REFRESH_ALL_FILES: 'refreshAllFiles',
-  UPDATE_INPUT_FILES: 'updateInputFiles',
-  UPDATE_REFERENCE_FILES: 'updateReferenceFiles',
-  UPDATE_AUXILIARY_FILES: 'updateAuxiliaryFiles',
-  UPDATE_MEDIA_FILES: 'updateMediaFiles',
-  UPDATE_OUTPUT_FILES: 'updateOutputFiles',
-
-  // Git/diff operations
-  REQUEST_RECENT_COMMITS: 'requestRecentCommits',
-  REFRESH_COMMITS: 'refreshCommits',
-  LATEXDIFF: 'latexdiff',
-  LATEXDIFFVC: 'latexdiffvc',
-  PACK_LATEXDIFFVC: 'packLatexdiffvc',
-  CLEAN_LATEXDIFFVC: 'cleanLatexdiffvc',
-
-  // Housekeeping operations
-  CLEAN_OUTPUT: 'cleanOutput',
-  CLEAN_BUILD: 'cleanBuild',
-  INDENT_TEX: 'indentTeX',
-  PACK_SINGLE: 'packSingle',
-  CLEAN_SINGLE: 'cleanSingle',
-  PACK_MULTIPLE: 'packMultiple',
-  CLEAN_MULTIPLE: 'cleanMultiple',
-
-  // Other operations
-  ACCEPT_EDITED: 'acceptEdited',
-
-  // Followup task (receives from progress view)
-  SETUP_FOLLOWUP_TASK: 'setupFollowupTask',
-} as const;
-
-// Progress view specific commands
-export const PROGRESS_VIEW_COMMANDS = {
-  ...COMMON_COMMANDS,
-  // Stream management
-  SWITCH_STREAM: 'switchStream',
-  DELETE_STREAM: 'deleteStream',
-  CLEAN_STREAM: 'cleanStream',
-  STOP_STREAM: 'stopStream',
-  UPDATE_STREAMS: 'updateStreams',
-  DELETE_ALL: 'deleteAll',
-
-  // Logging
-  LOG_DELTA: 'logDelta',
-
-  // Instruction panel
-  UPDATE_INSTRUCTION: 'updateInstruction',
-
-  // Todo List
-  UPDATE_TODOS: 'updateTodos',
-
-  // Queued follow-ups
-  UPDATE_QUEUED_FOLLOW_UPS: 'updateQueuedFollowUps',
-
-  // Batched metadata sync on tab switch (non-log content only)
-  SYNC_STREAM_CONTENT: 'syncStreamContent',
-
-  // Status and files
-  UPDATE_STREAM_STATUS: 'updateStreamStatus', // Update single stream's status in tabs
-  SET_ACTIVE_STREAM: 'setActiveStream',
-  UPDATE_CONVERSATION_PROGRESS: 'updateConversationProgress',
-  UPDATE_STREAM_BADGES: 'updateStreamBadges',
-  UPDATE_PARENT_STREAM: 'updateParentStream',
-  UPDATE_FILES: 'updateFiles',
-  UPDATE_MISSING_OUTPUTS: 'updateMissingOutputs',
-  UPDATE_PERMISSION: 'updatePermission',
-  UPDATE_BYPASS: 'updateBypass',
-
-  // Usage
-  UPDATE_RUN_USAGE: 'updateRunUsage', // Update single run's usage (incremental)
-
-  // Actions
-  RESUME: 'resume',
-  RUN_NEW: 'runNew',
-  COMPACT_RESPONSE: 'compactResponse',
-  RETRY_STREAM_REQUEST: 'retryStreamRequest',
-  CANCEL_RETRY_REQUEST: 'cancelRetryRequest',
-  DIFF_STREAM: 'diffStream',
-  PACK_STREAM: 'packStream',
-  SORT_STREAMS: 'sortStreams',
-  FILTER_STREAMS: 'filterStreams',
-  RESTORE_STATE: 'restoreState',
-  SEND_FOLLOW_UP: 'sendFollowUp',
-  POLISH_FOLLOW_UP: 'polishFollowUp',
-  UPDATE_FOLLOW_UP_TEXT: 'updateFollowUpText',
-  START_RECORDING: 'startRecording',
-  STOP_RECORDING: 'stopRecording',
-  UPDATE_RECORDING: 'updateRecording',
-  SHOW_INFORMATION_MESSAGE: 'showInformationMessage',
-  OPEN_TASK_STORAGE: 'openTaskStorage',
-  TOOL_EDIT_APPROVAL_ACTION: 'toolEditApprovalAction',
-  TOGGLE_TOOL_EDIT_APPROVAL_BYPASS: 'toggleToolEditApprovalBypass',
-  AGENT_PROPOSAL_ACTION: 'agentProposalAction',
-  BASH_APPROVAL_ACTION: 'bashApprovalAction',
-  RESTORE_PROPOSAL_CONFIG: 'restoreProposalConfig',
-  TOGGLE_SUPER_YOLO_BYPASS: 'toggleSuperYoloBypass',
-
-  // Memory
-  OPEN_MEMORY_VIEW: 'openMemoryView',
-
-  // File operations
-  OPEN_FILE: 'openFile',
-  OPEN_FILE_COMPILE: 'openFileCompile',
-  COMPARE_ORIGINAL: 'compareOriginal',
-  COMPARE_PREVIOUS: 'comparePrevious',
-  ACCEPT_FILE: 'acceptFile',
-  MERGE_FILE: 'mergeFile',
-  LATEXDIFF_FILE: 'latexdiffFile',
-  OPEN_LABEL: 'openLabel',
-
-  // Profile
-  OPEN_PROFILE: 'openProfile',
-
-  // Followup task (workflow continuation)
-  SETUP_FOLLOWUP: 'setupFollowup',
-  RUN_FOLLOWUP: 'runFollowup',
-  GET_FOLLOWUP_OPTIONS: 'getFollowupOptions',
-  SET_FOLLOWUP_OPTIONS: 'setFollowupOptions',
-  POP_OUT: 'popOut',
-  POP_BACK: 'popBack',
-  SET_PLACEMENT: 'setPlacement',
-} as const;
-
-// History view specific commands
-export const HISTORY_VIEW_COMMANDS = {
-  ...COMMON_COMMANDS,
-  GET_HISTORY_DATA: 'getHistoryData',
-  UPDATE_HISTORY: 'updateHistory',
-  CLEAR_HISTORY: 'clearHistory',
-  HISTORY_CLEARED: 'historyCleared',
-  RERUN_AGENT: 'rerunAgent',
-  RESTORE_AGENT: 'restoreAgent',
-  DELETE_AGENT: 'deleteAgent',
-} as const;
-
-// Profile view specific commands
-export const PROFILE_VIEW_COMMANDS = {
-  ...COMMON_COMMANDS,
-  GET_PROFILE_DATA: 'getProfileData',
-  UPDATE_PROFILE: 'updateProfile',
-  SELECT_AGENT: 'selectAgent',
-  SIGN_IN: 'signIn',
-  SIGN_OUT: 'signOut',
-  // API access mode toggle (Ultra tier)
-  SET_API_ACCESS_MODE: 'setApiAccessMode',
-} as const;
-
-// Memory view specific commands
-export const MEMORY_VIEW_COMMANDS = {
-  ...COMMON_COMMANDS,
-  GET_MEMORY_DATA: 'getMemoryData',
-  UPDATE_MEMORY: 'updateMemory',
-  OPEN_MEMORY_FILE: 'openMemoryFile',
-  OPEN_MEMORY_FOLDER: 'openMemoryFolder',
-  DELETE_MEMORY: 'deleteMemory',
-  GET_MEMORY_ENABLED: 'getMemoryEnabled',
-  SET_MEMORY_ENABLED: 'setMemoryEnabled',
-  UPDATE_MEMORY_ENABLED: 'updateMemoryEnabled',
-} as const;
-
-/**
- * Command string literals for settings view schema definitions.
- * Defined here (not in settingsViewMessages.ts) to avoid circular dependency:
- * commands.ts → settingsViewMessages.ts → memoryViewMessages.ts → commands.ts
- */
-export const SETTINGS_VIEW_CMD = {
-  // Navigation commands
-  SET_TAB: 'setTab',
-  OPEN_VSCODE_SETTINGS: 'openVscodeSettings',
-  // Memory commands
-  GET_MEMORY_DATA: 'getMemoryData',
-  OPEN_MEMORY_FILE: 'openMemoryFile',
-  OPEN_MEMORY_FOLDER: 'openMemoryFolder',
-  DELETE_MEMORY: 'deleteMemory',
-  GET_MEMORY_ENABLED: 'getMemoryEnabled',
-  SET_MEMORY_ENABLED: 'setMemoryEnabled',
-  // History commands
-  GET_HISTORY_DATA: 'getHistoryData',
-  RERUN_AGENT: 'rerunAgent',
-  RESTORE_AGENT: 'restoreAgent',
-  DELETE_AGENT: 'deleteAgent',
-  CLEAR_HISTORY: 'clearHistory',
-  // Profile commands
-  GET_PROFILE_DATA: 'getProfileData',
-  SELECT_AGENT: 'selectAgent',
-  SIGN_IN: 'signIn',
-  SIGN_OUT: 'signOut',
-  SET_API_ACCESS_MODE: 'setApiAccessMode',
-  SET_PROVIDER_KEY: 'setProviderKey',
-  REMOVE_PROVIDER_KEY: 'removeProviderKey',
-  OPEN_PROVIDER_KEY_URL: 'openProviderKeyUrl',
-  SET_PROVIDER_STREAMING: 'setProviderStreaming',
-  SET_PROVIDER_ENDPOINT: 'setProviderEndpoint',
-  SET_GLOBAL_STREAMING: 'setGlobalStreaming',
-  SET_PROVIDER_VSCODE_SETTING: 'setProviderVscodeSetting',
-  OPEN_EXTERNAL_URL: 'openExternalUrl',
-  // Model selection commands
-  GET_MODEL_SELECTION: 'getModelSelection',
-  SET_MODEL_ENABLED: 'setModelEnabled',
-  SET_HELPER_MODEL: 'setPolishModel',
-  // Agent selection commands
-  GET_AGENT_SELECTION: 'getAgentSelection',
-  OPEN_AGENT_YAML: 'openAgentYaml',
-  SET_AGENT_ENABLED: 'setAgentEnabled',
-  SET_ALL_AGENTS_ENABLED: 'setAllAgentsEnabled',
-  OPEN_AGENT_FOLDER: 'openAgentFolder',
-  CREATE_AGENT: 'createAgent',
-  CUSTOMIZE_AGENT: 'customizeAgent',
-  DELETE_CUSTOM_AGENT: 'deleteCustomAgent',
-  REVEAL_AGENT_FILE: 'revealAgentFile',
-  // Custom agent directory commands
-  GET_CUSTOM_AGENT_DIR: 'getCustomAgentDir',
-  SET_CUSTOM_AGENT_DIR: 'setCustomAgentDir',
-  RESET_CUSTOM_AGENT_DIR: 'resetCustomAgentDir',
-  // Multi-Agent commands
-  GET_SUPER_YOLO_ENABLED: 'getSuperYoloEnabled',
-  SET_SUPER_YOLO_ENABLED: 'setSuperYoloEnabled',
-  SET_ALLOW_ORCHESTRATOR_KILL: 'setAllowOrchestratorKill',
-  APPLY_AGENT_MODE_PRESET: 'applyAgentModePreset',
-  SAVE_AGENT_MODE_PRESET: 'saveAgentModePreset',
-  DELETE_AGENT_MODE_PRESET: 'deleteAgentModePreset',
-  GET_AGENT_MODE_PRESETS: 'getAgentModePresets',
-  // Tool dashboard commands
-  GET_TOOL_DASHBOARD_DATA: 'getToolDashboardData',
-  OPEN_TOOL_INSTALL_URL: 'openToolInstallUrl',
-  INSTALL_TOOL_EXTENSION: 'installToolExtension',
-  RECHECK_TOOL_STATUS: 'recheckToolStatus',
-  // LaTeX settings commands
-  GET_LATEX_SETTINGS_STATUS: 'getLatexSettingsStatus',
-  APPLY_LATEX_SETTINGS: 'applyLatexSettings',
-  INSTALL_LATEX_WORKSHOP: 'installLatexWorkshop',
-  RUN_INSTALL_COMMAND: 'runInstallCommand',
-} as const;
-
-// Settings view specific commands (combines Memory, History, and Profile views)
-// SETTINGS_VIEW_CMD is the source of truth; outbound-only commands are added here
-export const SETTINGS_VIEW_COMMANDS = {
-  ...COMMON_COMMANDS,
-  ...SETTINGS_VIEW_CMD,
-  // Outbound-only commands (backend → frontend, not schema-validated)
-  UPDATE_MEMORY: 'updateMemory',
-  UPDATE_MEMORY_ENABLED: 'updateMemoryEnabled',
-  UPDATE_HISTORY: 'updateHistory',
-  HISTORY_CLEARED: 'historyCleared',
-  UPDATE_PROFILE: 'updateProfile',
-  UPDATE_MODEL_SELECTION: 'updateModelSelection',
-  UPDATE_AGENT_SELECTION: 'updateAgentSelection',
-  UPDATE_CUSTOM_AGENT_DIR: 'updateCustomAgentDir',
-  UPDATE_SUPER_YOLO_ENABLED: 'updateSuperYoloEnabled',
-  UPDATE_AGENT_MODE_PRESETS: 'updateAgentModePresets',
-  UPDATE_TOOL_DASHBOARD: 'updateToolDashboard',
-  UPDATE_LATEX_SETTINGS_STATUS: 'updateLatexSettingsStatus',
-} as const;
+export { COMMON_COMMANDS } from './commonCommands';
+export { MAIN_VIEW_COMMANDS } from './mainViewCommands';
+export { PROGRESS_VIEW_COMMANDS } from './progressViewCommands';
+export { HISTORY_VIEW_COMMANDS } from './historyViewCommands';
+export { PROFILE_VIEW_COMMANDS } from './profileViewCommands';
+export { MEMORY_VIEW_COMMANDS } from './memoryViewCommands';
+export { SETTINGS_VIEW_CMD, SETTINGS_VIEW_COMMANDS } from './settingsViewCommands';
