+++ src/common/webview/historyViewCommands.ts
@@ -1,0 +1,16 @@
+/**
+ * Command constants for the history view.
+ */
+import { COMMON_COMMANDS } from './commonCommands';
+
+// History view specific commands
+export const HISTORY_VIEW_COMMANDS = {
+  ...COMMON_COMMANDS,
+  GET_HISTORY_DATA: 'getHistoryData',
+  UPDATE_HISTORY: 'updateHistory',
+  CLEAR_HISTORY: 'clearHistory',
+  HISTORY_CLEARED: 'historyCleared',
+  RERUN_AGENT: 'rerunAgent',
+  RESTORE_AGENT: 'restoreAgent',
+  DELETE_AGENT: 'deleteAgent',
+} as const;
