+++ src/webview/MainViewMessageHandler.ts
@@ -4,6 +4,7 @@ import {
   dispatchMainViewInbound,
   MainViewInboundHandlerRegistry,
 } from '@shared/schemas';
+import { PROVIDER_URLS } from '@shared/constants/providers';
 import { toErrorMessage } from '@common/errors';
 import {
   BaseViewMessageHandler,
@@ -18,7 +19,6 @@ import { showInstructionWithSuppress } from '@frontend/ui/instruction';
 import { getConfig, updateConfig, SETTINGS_QUERY } from '@utils/config';
 import { checkCoreDependencies, getToolDocsCommand } from '@utils/system';
 import { AUTH_COMMANDS, getAuthStatus } from '@commands/auth';
-import { PROVIDER_URLS } from '@commands/api/apiKeyCommands';
 
 import { DiffManager } from './managers/DiffManager';
 import { ExecutionManager } from './managers/ExecutionManager';
